% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysis.R
\name{mod_analysis_ui}
\alias{mod_analysis_ui}
\alias{mod_analysis_server}
\title{mod_analysis_ui and mod_analysis_server}
\usage{
mod_analysis_ui(id)

mod_analysis_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
