% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_biomass.R
\name{mod_biomass_ui}
\alias{mod_biomass_ui}
\alias{mod_biomass_server}
\title{mod_biomass_ui and mod_biomass_server}
\usage{
mod_biomass_ui(id)

mod_biomass_server(input, output, session, params)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
