% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fish.R
\name{mod_fish_ui}
\alias{mod_fish_ui}
\alias{mod_fish_server}
\title{mod_fish_ui and mod_fish_server}
\usage{
mod_fish_ui(id)

mod_fish_server(input, output, session, params)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
