% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_parameter_table.R
\name{mod_parameter_table_ui}
\alias{mod_parameter_table_ui}
\alias{mod_parameter_table_server}
\title{mod_parameter_table_ui and mod_parameter_table_server}
\usage{
mod_parameter_table_ui(id)

mod_parameter_table_server(input, output, session, params)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
