% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_parameters.R
\name{mod_parameters_ui}
\alias{mod_parameters_ui}
\alias{mod_parameters_server}
\title{mod_parameters_ui and mod_parameters_server}
\usage{
mod_parameters_ui(id)

mod_parameters_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
