% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_recruitment.R
\name{mod_recruitment_ui}
\alias{mod_recruitment_ui}
\alias{mod_recruitment_server}
\title{mod_recruitment_ui and mod_recruitment_server}
\usage{
mod_recruitment_ui(id)

mod_recruitment_server(input, output, session, params)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
