% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{mod_report_ui}
\alias{mod_report_ui}
\alias{mod_report_server}
\title{mod_report_ui and mod_report_server}
\usage{
mod_report_ui(id)

mod_report_server(input, output, session, params)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
