% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_schedule.R
\name{mod_schedule_ui}
\alias{mod_schedule_ui}
\alias{mod_schedule_server}
\title{mod_schedule_ui and mod_schedule_server}
\usage{
mod_schedule_ui(id)

mod_schedule_server(input, output, session, params)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
