% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_yield.R
\name{mod_yield_ui}
\alias{mod_yield_ui}
\alias{mod_yield_server}
\title{mod_yield_ui and mod_yield_server}
\usage{
mod_yield_ui(id)

mod_yield_server(input, output, session, params)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
