% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_ypr_app}
\alias{run_ypr_app}
\title{Run Shiny Application}
\usage{
run_ypr_app(population = NULL)
}
\arguments{
\item{population}{A ypr population object}
}
\description{
Run Shiny Application
}
\examples{
if(interactive()){run_ypr_app()}
}
