% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgtemplate-package.R
\docType{package}
\name{shinyypr-package}
\alias{shinyypr}
\alias{shinyypr-package}
\title{shinyypr: Run ypr Shiny App}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

An R package to run the ypr Shiny app.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/shinyypr}
  \item Report bugs at \url{https://github.com/poissonconsulting/shinyypr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Sebastian Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID}) [contributor]
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
